Rename: 폴더와 jsx 파일의 이름이 같은 경우 index로 파일명 변경
